Crate workflow_visualizer

Workflow Visualizer is a lib for cross-platform declarative UI It is comprised of three major parts. trait Workflow to setup application messaging. Visualizer which handles rendering Runner which invokes the visualizer’s tools for multiple platforms

Re-exports

pub use bevy_ecs;

pub use wgpu;

pub use winit;

Structs

Area

Area is for width/height of an entity

Attachment

Used to hold queued attachments until ready to invoke attach to the Visualizer

AttributeWrite

Temporary storage for staging writes to the Gpu

Color

RGBA colors

ContentLocation

Pair of ContentMarker and Offset to get an exact grid location

ContentMarker

Logical index using groupings to get actual markers then to px

ContentRange

Beginning and End GridLocation grouping

ContentView

A GridRange for horizontal + vertical aspects

Coordinate

Coordinate is a bundle of Section + Layer to denote coordinates in the world

CpuAttributeBuffer

Cpu side buffer for an attribute of instances

DeviceContext

Device Context is used for actual physical positions such as hardware sizes

EnableVisibility

Entry point for enabling visibility module on an entity

EntityName

Focus

Used to set the Focus of an element

FocusInputListener

used to opt into VirtualKeyboard open functionality

GfxOptions

Options for instantiating the Gfx module

GfxSurface

Core structs for interop with wgpu

GpuAttributeBuffer

Gpu side buffer of an attribute of an instance

Grid

Index

order in the instance buffers for a specific entity

Indexer

Manager for indexes that increments and holds the current count of instances

InstanceAttributeManager

Combines a cpu/gpu/write for an attribute for full management of the data for an attribute

InterfaceContext

Interface Context is used for logical coordinates after accounting for scale factor

Job

Key

Key to associate with an index for consistent referencing

KeyFactory

generator for keys that are unique within a factory

Layer

Layer represents what plane this entity resides on. Used to differentiate z coords in rendering.

Marker

Index of 8px alignment location

MarkerGrouping

Number of markers to include in a logical group

NullBit

NullBit is useful for dropping instances without having to resize the buffer or swap elements. NullBit sets all other attributes to 0 in the shader when multiplied by to cancel one instance while still having it included in the buffer.

NumericalContext

Numerical Context is used for numbers that are not associated with device/logical sizes

Position

Position denotes 2d coordinates in space with float32 precision

RawArea

Raw area defined in C representation for interacting with C (vulkan mostly)

RawPosition

Raw position for interacting with C

RenderPassHandle

Request

ResponsiveView

A mapping of GridView for each HorizontalSpan Option

Runner

Main struct to run the visualizer’s event loop

ScaleFactor

Section

A section is a bundle for Position/Area

Sender

Sender is for sending actions to the app from within the visualizer

Text

The text to render

TextBundle

Entry point to spawn a Text element

TextGridLocation

where the letter is in the Text Grid

TextGridPlacement

Mapping of what glyphs keys are placed where in the TextGrid

TextLetterDimensions

The size of a letter at an alignment

TextLineStructure

Lengths of the lines present in the Text

TextScale

The scale of the Text

TextWrapStyle

Whether to wrap by Word or Letter

Theme

ThemeDescriptor

TimeDelta

signifies a change in time

TimeMarker

signifies a point in time

Timer

Access to time on the platform

TouchListener

Touchable

Uniform

Viewport

ViewportAttachment

ViewportHandle

Visibility

Current Visibility

VisibleSection

What part of the entity is visible

Visualizer

Struct for preparing/configuring/rendering visuals

WindowAttachment

WindowResize

Enums

ContentOffset

whether to attach to beginning/end of column

JobSyncPoint

RenderPhase

SyncPoint

TextScaleAlignment

What predetermined size category to assign to the Text

UserSpaceSyncPoint

Traits

Attach

Trait for configuring how a part attaches to the Job of a visualizer

Render

Workflow

Main trait to establish communication between the app and UI thread.

Functions

spawn

start_web_worker

spawn a web worker using the types blanket implemented gloo_worker

Type Definitions

ResponsiveContentView