Struct workflow_visualizer::Section

pub struct Section<Context: CoordinateContext> {
    pub position: Position<Context>,
    pub area: Area<Context>,
}

A section is a bundle for Position/Area

Fields
position: Position<Context>area: Area<Context>
Implementations

impl<Context: CoordinateContext> Section<Context>

pub fn new<P: Into<Position<Context>>, A: Into<Area<Context>>>( position: P, area: A ) -> Self

pub fn from_left_top_right_bottom( left: f32, top: f32, right: f32, bottom: f32 ) -> Self

Can be instantiated with specific points

pub fn width(&self) -> f32

pub fn height(&self) -> f32

pub fn left(&self) -> f32

pub fn right(&self) -> f32

pub fn top(&self) -> f32

pub fn bottom(&self) -> f32

pub fn is_touching(&self, other: Self) -> bool

returns if any port of this section is touching the other

pub fn is_overlapping(&self, other: Self) -> bool

returns true if section overlaps the other

pub fn contains(&self, position: Position<Context>) -> bool

returns true if the position resides in the section

pub fn intersection(&self, other: Self) -> Option<Self>

returns an Option of the overlap between the sections