Crate workflow_visualizer

Re-exports

pub use async_trait;

pub use bevy_ecs;

pub use wgpu;

pub use winit;

Structs

Area

Area base coordinate class for 2d area requires setting a context to differentiate and track what the area means

Attachment

AttributeWrite

Color

Coordinate

Coordinate is a bundle of Section + Layer to denote coordinates in the world

CpuAttributeBuffer

DeviceContext

Device Context is used for actual physical positions such as hardware sizes

EnableVisibility

EntityName

Focus

FocusInputListener

GfxOptions

GfxSurface

GpuAttributeBuffer

Index

Indexer

InstanceAttributeManager

InterfaceContext

Interface Context is used for logical coordinates after accounting for scale factor

Job

Key

KeyFactory

Layer

Layer represents what plane this entity resides on. Used to differentiate z coords in rendering.

NullBit

NumericalContext

Numerical Context is used for numbers that are not associated with device/logical sizes

Position

Position denotes 2d coordinates in space with float32 precision

RawArea

Raw area defined in C representation for interacting with C (vulkan mostly)

RawPosition

Raw position for interacting with C

RenderPassHandle

Request

Runner

ScaleFactor

Section

Sender

Text

TextGridLocation

TextGridPlacement

TextLetterDimensions

TextLineStructure

TextRequest

TextScale

TextWrapStyle

Theme

ThemeDescriptor

TimeDelta

TimeMarker

Timer

TouchListener

Touchable

Uniform

Viewport

ViewportAttachment

ViewportHandle

Visibility

VisibleSection

Visualizer

WindowAttachment

WindowResize

Enums

JobSyncPoint

RenderPhase

SyncPoint

TextScaleAlignment

UserSpaceSyncPoint

Traits

Attach

Render

Workflow

Functions

spawn

start_web_worker