Crate workflow_visualizer

Workflow Visualizer is a lib for cross-platform declarative UI It is comprised of three major parts. trait Workflow to setup application messaging. Visualizer which handles rendering Runner which invokes the visualizer’s tools for multiple platforms

Re-exports

pub use bevy_ecs;

pub use wgpu;

pub use winit;

Structs

ActiveInteraction

Animation

Area

Area is for width/height of an entity

AtlasBindGroup

AtlasBlock

AtlasDimension

AtlasFreeLocations

AtlasPosition

AtlasTexture

AtlasTextureDimensions

Attachment

Used to hold queued attachments until ready to invoke attach to the Visualizer

AttributeWrite

Temporary storage for staging writes to the Gpu

BackgroundColor

BorderColor

BundleBuilder

Button

ButtonDespawn

Color

RGBA colors

Coordinate

Coordinate is a bundle of Section + Layer to denote coordinates in the world

CpuAttributeBuffer

Cpu side buffer for an attribute of instances

DeviceContext

Device Context is used for actual physical positions such as hardware sizes

Diagnostics

Interpreter for Diagnostic traces

DiagnosticsHandle

Handle for use in Local<…> resources in systems to record info

Disabled

EnableVisibility

Entry point for enabling visibility module on an entity

Exit

Exit hook

Focus

Used to set the Focus of an element

FocusInputListener

used to opt into VirtualKeyboard open functionality

GfxOptions

Options for instantiating the Gfx module

GfxSurface

Core structs for interop with wgpu

GfxSurfaceConfiguration

GpuAttributeBuffer

Gpu side buffer of an attribute of an instance

Grid

Grid configuration of the Span + Column/Row/Gutter Configs

GridLocation

Description of a Location on the Grid

GridLocationDescriptor

Pair of GridMarker and Bias to get an exact grid location

GridMarker

Logical index using groupings to get actual markers then to px

GridPoint

Point in a Grid with x/y as GridLocations

GridRange

Beginning and End GridLocation grouping

GridView

A GridRange for horizontal + vertical aspects

Icon

IconBitmap

IconBitmapRequest

IconId

IconPixelData

Idle

Idle hook

Index

order in the instance buffers for a specific entity

Indexer

Manager for indexes that increments and holds the current count of instances

InstanceAttributeManager

Combines a cpu/gpu/write for an attribute for full management of the data for an attribute

Interactable

Interaction

InteractionEvent

InteractionLocation

InteractionLocations

InteractionPhases

InteractionTracker

InterfaceContext

Interface Context is used for logical coordinates after accounting for scale factor

Interpolation

InterpolationExtraction

Job

Extensible container + task runner

Key

Key to associate with an index for consistent referencing

KeyFactory

generator for keys that are unique within a factory

Layer

Layer represents what plane this entity resides on. Used to differentiate z coords in rendering.

Line

LineRender

MsaaRenderAdapter

Used for setting the multi-sampling/anti-aliasing

NullBit

NullBit is useful for dropping instances without having to resize the buffer or swap elements. NullBit sets all other attributes to 0 in the shader when multiplied by to cancel one instance while still having it included in the buffer.

NumericalContext

Numerical Context is used for numbers that are not associated with device/logical sizes

Panel

PanelContentArea

Path

Collection of specific points rendered from a PathView

PathView

Collection of PathViewPoints

PlacementReference

Position

Position denotes 2d coordinates in space with float32 precision

PrimaryInteraction

PrimaryMouseButton

RawArea

Raw area defined in C representation for interacting with C (vulkan mostly)

RawMarker

Index of 8px alignment location

RawMarkerGrouping

Number of markers to include in a logical group

RawPosition

Raw position for interacting with C

RenderPassHandle

Wrapper around wgpu::RenderPass

ResponsiveView

A mapping of GridView for each HorizontalSpan Option

Runner

Main struct to run the visualizer’s event loop

ScaleFactor

Scale Factor of the device’s Window

Section

A section is a bundle for Position/Area

Sender

Sender is for sending actions to the app from within the visualizer

Tag

Text

Entry point to spawn a Text element

TextGridLocation

where the letter is in the Text Grid

TextGridPlacement

Mapping of what glyphs keys are placed where in the TextGrid

TextLetterDimensions

The size of a letter at an alignment This holds Area<DeviceContext> as character dimensions are scaled before rasterization to prevent visual artifacts of stretching

TextLineStructure

Lengths of the lines present in the Text

TextScale

The scale of the Text

TextValue

The text to render

TextWrapStyle

Whether to wrap by Word or Letter

TextureAtlas

TextureCoordinates

Theme

Color definitions for element types in the Visualizer

ThemeDescriptor

Describes a Theme

TimeDelta

signifies a change in time

TimeMarker

signifies a point in time

Timer

Access to time on the platform

Toggled

Triggered

Uniform

Wrapper around wgpu::Buffer for use as a Uniform Buffer

Viewport

Viewport Matrix for converting to NDC

ViewportHandle

Handle to a Viewport to adjust position/get section

Visibility

Current Visibility

VisibleSection

What part of the entity is visible

Visualizer

Struct for preparing/configuring/rendering visuals

WindowResize

Event for triggering Window Resizing behaviour

Enums

BundledIcon

ButtonBorder

ButtonType

GridMarkerBias

Whether to attach to beginning/end of column

HorizontalSpan

Span used for setting the number of columns available in the Grid

IconScale

InteractionPhase

JobSyncPoint

SyncPoint for Job Idle behaviour

PanelType

RenderPhase

Phase for Rendering

SyncPoint

Synchronization Points for bucketing systems in a task

TextScaleAlignment

What predetermined size category to assign to the Text

Traits

Animate

Attach

Trait for configuring how a part attaches to the Job of a visualizer

BundleExtension

BundlePlacement

Record

How to record aspects to your recorder

Render

Trait to extend the render loop with a new pipeline

ResponsiveUnit

Shorthand for specifying a GridLocation using near/far bias

Workflow

Main trait to establish communication between the app and UI thread.

Functions

start_web_worker

spawn a web worker using the types blanket implemented gloo_worker

Type Definitions

ButtonTag

IconTag

LineTag

PanelTag

ResponsiveGridPoint

Convenience type for ResponsiveView<GridPoint>;

ResponsiveGridView

Convenience type for mapping to ContentViews

ResponsivePathView

Responsive Mapping for PathView

TextTag